reorganized the sites directory
