+++ job/steps/scrape_metadata.py
@@ -15,7 +15,10 @@ from db.mappings.base import db_proxy
 from db.mappings.path import Path
 from job.steps import warehouse
 from lib.metrics import Unit, write_metric
-from sites.helpers import ArticleScrapingError, ScrapeFailure
+
+from sites.article_scraping_error import (
+    ArticleScrapingError,
+    ScrapeFailure)
 from sites.site import Site
 
 EXCLUDE_FAILURE_TYPES = {
