+++ sites/philadelphia_inquirer.py
@@ -5,13 +5,20 @@ from typing import Any, Dict, List, Optional, Union
 from requests.models import Response
 
 from lib.config import config
-from sites.helpers import (
-    GOOGLE_TAG_MANAGER_RAW_FIELDS,
-    ArticleScrapingError,
-    ScrapeFailure,
-    ms_timestamp,
-    safe_get,
+from sites.ms_timestamp import ms_timestamp
+
+from sites.google_tag_manager import (
     transform_data_google_tag_manager,
+    GOOGLE_TAG_MANAGER_RAW_FIELDS
+)
+
+from sites.article_scraping_error import (
+    ArticleScrapingError,
+    ScrapeFailure)
+
+from sites.safe_get import safe_get
+
+from sites.validate import (
     validate_response,
     validate_status_code,
 )
