+++ tests/sites/test_helpers.py
@@ -1,7 +1,10 @@
 import pytest
 from requests.models import Response
 
-from sites.helpers import validate_response, validate_status_code
+from sites.validate import (
+    validate_response,
+    validate_status_code,
+)
 
 
 def _create_response(status_code: int) -> Response:
